requirements.txt
customtkinter
pymupdf
pypdf
pillow
pyinstaller
